refactor: :recycle: Refactor into optional deps
