+++ pyproject.toml
@@ -5,11 +5,9 @@ description = "A framework for building autonomous AI agents"
 readme = "README.md"
 requires-python = ">=3.11"
 dependencies = [
-    "anthropic>=0.37.1",
     "bs4>=0.0.2",
     "docstring-parser>=0.16",
     "httpx>=0.27.2",
-    "openai>=1.52.0",
     "python-dotenv>=1.0.1",
     "rich>=13.9.2",
     "typer>=0.12.5",
@@ -18,9 +16,14 @@ dependencies = [
 [tool.uv]
 package = true
 override-dependencies = ["greenlet==3.1.0"]
+dev-dependencies = ["pytest>=8.3.3"]
 
 [project.scripts]
 tono = "tono.cli:app"
+
+[project.optional-dependencies]
+openai = ["openai>=1.52.0"]
+anthropic = ["anthropic>=0.37.1"]
 
 [build-system]
 requires = ["hatchling"]
