+++ src/tono/__init__.py
@@ -1,5 +1,7 @@
-from tono.agent import Agent  # noqa
-from tono.ai import OpenAICompletionClient, AnthropicCompletionClient  # noqa
+from tono.lib.agent import Agent  # noqa
 
 __app_name__ = "tono"
 __version__ = "0.1.1"
+
+
+__all__ = ["Agent"]
