+++ src/tono/models/anthropic/__init__.py
@@ -1,0 +1,3 @@
+from tono.models.anthropic._client import CompletionClient  # noqa
+
+__all__ = ["CompletionClient"]
