+++ src/tono/models/openai/__init__.py
@@ -1,0 +1,3 @@
+from tono.models.openai._client import CompletionClient  # noqa
+
+__all__ = ["CompletionClient"]
